+++ whale-line-segments.el
@@ -136,20 +136,35 @@ icon name and the face.")
 
 ;;;; -- Position
 
+(defun wls--position--percent ()
+  "Get the percent position."
+  `(:propertize ("" mode-line-percent-position)
+                face whale-line-shadow))
+
+(defun wls--position--line-and-column ()
+  "Get the line and column."
+  (when-let* ((format (car-safe mode-line-position-column-line-format)))
+
+    `(:propertize ,(string-trim format) face whale-line-shadow)))
+
+(defun wls--position--doc-view ()
+  "Get the position in a document."
+  '((:propertize (:eval (format "%d/%d" (image-mode-window-get 'page) (doc-view-last-page-number)))
+               face whale-line-shadow)))
+
+(defun wls--position--default ()
+  "Render the default position segment."
+  `(,(wls--position--line-and-column)
+    ,(whale-line--spacer)
+    ,(wls--position--percent)))
+
 (defun wls--position ()
-  "Render position segment."
-  (let ((str (cond
-              ((eq major-mode 'doc-view-mode)
-               (concat
-                (number-to-string (image-mode-window-get 'page))
-                "/"
-                (number-to-string (doc-view-last-page-number))))
-
-              ((bound-and-true-p follow-mode)
-               "f: %l:%c %p%")
-
-              (t "%l:%c %p%"))))
-    `((:propertize ,str face whale-line-shadow))))
+  "Render the position segment."
+  (cond
+   ((eq major-mode 'doc-view-mode)
+    (wls--position--doc-view))
+   (t
+    (wls--position--default))))
 
 (whale-line-create-stateless-segment position
   :getter wls--position